requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psutil>=5.9.0
scapy>=2.5.0
pyyaml>=6.0
pystray>=0.19.0
pillow>=10.0.0
plyer>=2.1.0
pywebview>=4.4.0
pyinstaller>=6.0.0
